threat_intel/requirements.txt
flask==3.0.3
requests==2.31.0
python-whois==0.9.4
geoip2==4.8.0
python-dateutil==2.9.0
